4.6.1

+++ updater.php
@@ -2,10 +2,13 @@
 
 /** @var CUpdater $updater */
 if ($updater && $updater instanceof CUpdater) {
-
     if (!function_exists('sprint_migration_rmdir')) {
         function sprint_migration_rmdir($dir)
         {
+            if (!is_dir($dir)) {
+                return false;
+            }
+
             $files = array_diff(scandir($dir), ['.', '..']);
             foreach ($files as $file) {
                 (is_dir("$dir/$file")) ? sprint_migration_rmdir("$dir/$file") : unlink("$dir/$file");
@@ -15,13 +18,12 @@ if ($updater && $updater instanceof CUpdater) {
     }
 
     if (!empty($_SERVER['DOCUMENT_ROOT'])) {
-//        sprint_migration_rmdir($_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/sprint.migration/lib/helpers/useroptions/');
-//        sprint_migration_rmdir($_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/sprint.migration/admin/');
-//        unlink($_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/sprint.migration/loader.php');
+        //        sprint_migration_rmdir($_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/sprint.migration/lib/helpers/useroptions/');
+        //        sprint_migration_rmdir($_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/sprint.migration/admin/');
+        //        unlink($_SERVER['DOCUMENT_ROOT'] . '/bitrix/modules/sprint.migration/loader.php');
     }
 
     if (is_dir(__DIR__ . '/install/gadgets/')) {
         $updater->CopyFiles("install/gadgets/", "gadgets/");
     }
-
 }
